(int)-1 != (short)-1. Yay for automatic integer type widening.
git-svn-id: svn+ssh://src.earth.threerings.net/vilya/trunk@324 c613c5cb-e716-0410-b11b-feb51c14d237

+++ src/java/com/threerings/whirled/spot/server/SpotSceneManager.java
@@ -190,7 +190,7 @@ public class SpotSceneManager extends SceneManager
     {
         Portal from = _enterers.remove(body.getOid());
         Portal entry;
-        if (from != null && from.targetPortalId != -1) {
+        if (from != null && from.targetPortalId != (short)-1) {
             entry = _sscene.getPortal(from.targetPortalId);
             if (entry == null) {
                 Log.warning("Body mapped at invalid portal [where=" + where() +
